agulbra-make-member inserted header file at the wrong place in case of a new file. This part of the function has now been rewritten to use part of klaralv.el, which has support for more situation anyway.
The support is only activated if the user has loaded klaralv.el, This is done by default in the
do-it-all setup by kde-emacs.el. So if klaralv.el is not loaded, then the user knows better, and then
we do not want to mess with him.

I take it for granted that you will bug me if there is any problems with my changes ;-)

svn path=/trunk/kdesdk/scripts/; revision=202179