Based on an observation by Frans Englich, remove the vast majority of the restrictions that kdesvn-build places on global options.
Now it is possible to override things like source-dir, qtdir, kdedir, and
svn-server on a *per-module* basis.  If you do this and break your install,
please don't blame it on me. =D  However it should make it possible to do
some cool things.

In addition, don't try to grab a lockfile when being run in --pretend mode. So
you can always run kdesvn-build -p even when kdesvn-build is already running.

svn path=/trunk/KDE/kdesdk/scripts/; revision=457211