This commit adds the ability to build and install languages from the l10n modules automatically.
It is controlled by the global kde-languages option, as described in kdesvn-buildrc-sample.

e.g.
global
  kde-languages el # greek
  ....
end global

If I remember right you can also control this via the l10n module directly (using
module l10n with checkout-only) but please don't. :)

Internally this is implemented in terms of the l10n module, so if you just want to update
your translations, it is as simple as:

% kdesvn-build l10n

or 

% kdesvn-build l10n/$LANG

PLEASE PLEASE PLEASE test this so I can make a release later. :)
In my testing l10n/de broke but l10n/fr worked (but that was last week).

I'd like to thank levipenumbra for providing an initial patch for the support.

Also included because I am lazy are some pretend fixes and a debugging chdir()
call.

svn path=/trunk/KDE/kdesdk/scripts/; revision=484473