Fix a rather longstanding and annoying issue with qt-copy 4 support.  kdesvn-build now adds the -prefix option to the qt-copy configure script and runs make install for qt 4.  Qt 3 is still fully supported, except that use-qt-builddir-hack (which makes kdesvn-build force qt-copy to be builddir != srcdir) is no longer optional (it is always enabled).
As always with any sort of change to the dance kdesvn-build plays with qt-copy, PLEASE TEST.

I even compiled (and installed for qt 4) both 3.3 and trunk qt-copy and it seems to be
alright.  I do still need to test the case where qtdir == $srcdir/build/qt-copy though.

Also, configure flags are no longer ignored for the l10n module.

BUG:116621

svn path=/trunk/KDE/kdesdk/scripts/; revision=518746