Script to replace QIODevice::at by either QIODevice::pos() or QIODevice::seek(offset), depending on whether at had an argument or not.
CAUTION: This script replaces all calls to methods named 'at', even if the object is not a QIODevice. So check whether the file contains only QIODevice calls or also QList etc. Alternatively, I usually apply the script only to files that throw a deprecated warning and compare the # of deprecated warnings to the nr of replacements.


svn path=/trunk/KDE/kdesdk/scripts/; revision=534521