Commit first phase of implementing wish 109619 (Perform svn update and make simultaneously).
This uses SysV interprocess communication (threads suck in Perl).  See your handy
dandy msgget(2) (and etc.) manpages if you're interested.  If you have svipc(7) that
might also be handy.

This still needs a lot of polishing, as both procs will try to output to screen at the
same time which makes the progress indicator hard to read.  Also handle_build and
handle_updates (but especially handle_build) haven't seen much love lately so I need
to clean those up a bit.

This is disabled for Coverity and in case the IPC comms line is unable to be setup, in
which case kdesvn-build will act much as it used to for an outside observer.

One problem is that using Ctrl-C to stop the script will leak the shared message queue
still at this point.  I think I can get an easy fix for that though.  If this happens
to you (too many leaked msg queues) consult the ipcs and ipcrm command line tools.

On a more depressing note, kdesvn-build is now 185K in size, uncompressed.  Yikes.

CCBUG:109619

svn path=/trunk/KDE/kdesdk/scripts/; revision=731251