Fix bug 151971 (kdesvn-build SVN aborts with IPC failure).
I was never able to pin down what was causing the bug, or to reliably reproduce
it.  So I went and completely ditched the old IPC code instead. :(

It has been rewritten using anonymous pipes and adding an extra process so that the
update and build processes can both proceed unfettered.

Also added is a command line option (--no-async) to disable the IPC stuff.  You can
set it as a config file option as well (just called async), it defaults to enabled.

In addition the async mode is turned off unless performing both the update and build
processes.  In theory async mode should always work but I see no reason to test the
issue unnecessarily.

BUG:151971

svn path=/trunk/KDE/kdesdk/scripts/; revision=735914