updates to svn{back,forward}port: more "standard" usage information, more portable/liberal detection of -h/--help, branch specification (backport, optional, default still 4.0) or read-from-actual-url (forwardport), proper argument quoting / use of "$@" (hopefully now safe for files with spaces), be more picky about what is "yes" when asking to commit, other minor stuff...
NOTE: POSIX states that 'test' with more than three arguments is not portable, and it's right; the old code could do something like 'test -z -gt -o -gt = -h -o -gt = --help', which is an error; the correct way (as stated many times on bug-bash) is to use && and ||, not test's -a/-o. Also use '[', as it looks nicer and I'm not aware of any reasonable shell these days that lacks [ as a built-in (even Solaris /bin/sh, which is *not* POSIX-compliant, has it), and have used it in my own must-be-portable scripts for ages.

svn path=/trunk/KDE/kdesdk/scripts/; revision=780733