Add persistent data store for kdesvn-build.  Initially the only things tracked are the last successful build revision, last successful install revision, and the number of consecutive module failures.
In addition the user is warned at program end of all modules that have failed to build for more
than 3 consecutive build attempts for the module to take a closer look at the issue.

If there's other cool things that would be useful for the script to track I'm all ears.  The
last-successful-rev stuff was initially for tracking if a module should be rebuilt or not.  But
if required library dependencies have changed in the interim (i.e. qt-copy or kdelibs) then we
should allow CMake to do its things no matter what so I will not actually implement that feature
after all.  But it's still useful to track IMO.

CCBUG:157992

svn path=/trunk/KDE/kdesdk/scripts/; revision=841748