Alright, I think this is a neat change.
kdesvn-build has supported an automatically generated symlink in the log
directory called latest for awhile now, which puts you in the log directory
for the last build run.

Now, instead of symlinking latest to the last build run, latest holds a
directory of symlinks.  Each symlink is named after the module concerned and
links back to the last build run for that module.  That way you always have
easy access to the logs for the last build run of a module without having to
run find or anything crazy.

You should find that using it is effectively the same unless you have tools
of your own that depend on $logdir/latest being a symlink.

svn path=/trunk/KDE/kdesdk/scripts/; revision=854548