Fix bug 172719 (cmake-options causes kdesvn-build to fail).
The root cause of the problem was that when you had more than one option on a single cmake-options
command that it was not getting split into a list properly.

The root cause of *that* error was that the regular expression I used to perform the split
(without splitting around double-quotes) failed with leading whitespace.  I didn't do testing for
that case but that it how the function ends up getting called in kdesvn-build.

In addition to the bug I've added quotes around individual command parameters for log_command() to
make future instances of this type of bug (improper splitting) more obvious.

BUG:172719

svn path=/trunk/KDE/kdesdk/scripts/; revision=910294