Improve git updating.
- git clone correctly sets up the __kdesvn-build-remote alias
- updating a module should hopefully "always work" since kdesvn-build always
  runs the git fetch and git checkout separately, so in theory (and from my
  experimentation) the result should be the same whether or not you were on
  the right repository already, and whether a remote tracking branch already
  existed or not.

One flaw is that if there is a git branch with the same name as the branch name
given in the kdesvn-buildrc file, then kdesvn-build will always use that one
(even if it doesn't point to the correct remote branch).  Not sure if this will
be an issue in practice though.

svn path=/trunk/KDE/kdesdk/scripts/; revision=1039964