Prefer git-pull to git-merge, and use git-stash if needed.
Specifically we prefer git pull --rebase (as recommended by dfaure, but jhbuild also
does this now apparently, plus it /is/ a good idea for most), as opposed to git merge.

We still need to do git fetch and git checkout first to setup the remote branches for a
given remote, and then setup the local branch for the remote branch. Oy.

In addition kdesrc-build will try to see if you have local changes, staged or not, and
will run git-stash for you first if that is the case, and then automatically run git-stash
pop afterwards.

Finally, while testing this I noticed that "xx files affected" messages didn't get printed
out when using --src-only (since that's technically printed out in the build phase now due
to async), so that is now fixed as well.

svn path=/trunk/KDE/kdesdk/scripts/; revision=1210586