kdesrc-build has moved to git. Delete everything not essential to kdesrc-build running, and in addition put in the most annoying exit-program message I can think of to nag people to cut their ties.
This should avoid confusing scripty while still warning users, without having to leave a
README.

I'd like to thank toma and the rest of the sysadmins for the quick turnaround on the
git.kde.org request, and the i18n team and ReviewBoard maintainer in advance for the
updates to those pieces of KDE infrastructure.

I plan on removing the script from svn entirely around 7 February or so.

CCBUG:263939

svn path=/trunk/KDE/kdesdk/doc/scripts/; revision=1216401