kde-emacs: Make whitespace removal optional.
Sometimes one is working on a one-liner and does not want to fix
whitespaces in the whole file, so it is now possible to set
'kde-emacs-delete-trailing-whitespace to nil to disable that.

GNU Emacs already has the delete-trailing-whitespace function; we can
hopefully remove the whitespace removal altogether when an XEmacs
equivalent shows up.

Reviewed-by: dfaure via email.


svn path=/trunk/KDE/kdesdk/scripts/; revision=1218721