Ensure that appPo and appDocs are always arrays.
The first time that appdata["app{Po,Docs}"] gets set, a simple string is
assigned, but it is apparently supposed to be an array of possible docs or
translations to add.

This didn't cause errors with Ruby 1.8 as the way the array was iterated also
happens to work on strings. Ruby 1.9 forces you to use each_line() on Strings
(or each_byte, each_character, etc.) and therefore fails which is how the
issue was caught.

BUG:267696

svn path=/trunk/KDE/kdesdk/scripts/; revision=1227514