Add CMake buildsystem for kde-dev-scripts.
I also removed some trailing whitespace while I'm at it. The buildsystem should allow
for installing to the appropriate directory even if KDE 4 is not present, but KDE 4
directories are used if they can be found.

This submodule was already optional in the top-level kdesdk so no change was required
there.

CCMAIL:kde-sdk-devel@kde.org

svn path=/trunk/KDE/kdesdk/scripts/; revision=1298173