When a brace is inserted for a C++ lambda, insert ; after close-brace
The emacs lisp function insert-curly-brace in kde-emacs changes the
behavior of inserting a '{' in C-mode to also insert a matching
'}'. This works well when definitng methods, writing if and for
statements etc. However, with C++1x lambdas, there needs to be a ;
following the definition of the lambda. This patch incorporates that
behavior by matching lambda syntax.

REQUIRES more TESTING.