kde-devel-gdb: fix printing QString
The old code long time didn't work: "len" field does not exist, data is
a method not a field, "unicode" does not exist, field "ucs" is not
present in data.

Let's just replace it with a new implementation. Credits for that goes
to user apouche from StackOverflow, whose implementation from 2012 year
was only hit by "data" being nowadays a method.

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>